separate router from UI
